+++ AGMPowerCLIGetFunctions.ps1
@@ -2623,5 +2623,3 @@ function Get-AGMWorkFlow ([string]$filtervalue,[string]$keyword,[switch][alias("
         Get-AGMAPIData -endpoint /workflow -limit $limit -sort $sort
     }
 }
-
-